requirements.txt
fastapi==0.95.1
uvicorn==0.18.3
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.1
joblib==1.2.0
psycopg2==2.9.6
kafka-python==2.0.2
websocket-client==1.5.1
pydantic==1.10.7
psycopg2-binary